Implement security middleware
